Correcting not escaped minus signs in manpage.
Extract of the explanations why we need a change for this:
By default, "-" chars are interpreted as hyphens (U+2010) by groff, not as minus
signs (U+002D). Since options to programs use minus signs (U+002D), this means
for example in UTF-8 locales that you cannot cut and paste options, nor search
for them easily. The Debian groff package currently forces "-" to be interpreted
as a minus sign due to the number of manual pages with this problem, but this is
a Debian-specific modification and hopefully eventually can be removed.

"-" must be escaped ("\-") to be interpreted as minus.